Improve NetTx exiting log (#1060)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1060

Currently, if there are too many seqs in `Outbox` or `Unacked`:

https://www.internalfb.com/code/fbsource/[315bec9cb3e7f40c08421ccefd1dfd489c3c55d3]/fbcode/monarch/hyperactor/src/channel/net.rs?lines=225%2C230%2C319-320

Their `Debug` string will be too long.

This diff wraps `MessageDeque` with a new type, and provide a custom Debug impl for that new type, in order to improve the debug string output.

In addition, this diff also adds a `sent_at` field to `struct QueuedMessage` so we can log when the message was written to the stream in the log.

A example log with this change will look like:

> 2025-09-02T13:49:18.076002Z DEBUG hyperactor::channel::net: session tcp:[::]:40871.3284287747997958583: NetTx exited its loop with state: Closing { deliveries: Deliveries { outbox: Outbox { next_seq: 318611, deque: [(seq=318610, since_rcv=14sec)] }, unacked: Unacked { deque: [(seq=316000, since_rcv=15sec, since_sent=0sec), (seq=316001, since_rcv=15sec, since_sent=0sec), (seq=316002, since_rcv=15sec, since_sent=0sec), ... omit 2604 messages ..., (seq=318607, since_rcv=14sec, since_sent=0sec), (seq=318608, since_rcv=14sec, since_sent=0sec), (seq=318609, since_rcv=14sec, since_sent=0sec)], largest_acked: Some((seq=315999, since_acked=0sec)) } }, reason: "session tcp:[::]:40871.3284287747997958583: failed to receive ack within timeout 15 secs; link is currently connected" }

Differential Revision: D81486270