mesh: introduce ValueMesh, ProcMeshRef::status
Summary:
ValueMeshes let us represent a mesh of values. We subsequently use this to represent the status of a proc mesh.

This ought to be made more generic. Ideally it would behave like an ordinary Rust collection (with the attendant implementations of FromIterator, etc.). The challenge presented is that not all iterators are valid ValueMeshes, since we require point compatibility and completeness. We also can't define a FromIterator implementation for Result<ValueMesh<T>, Error> because of the foreign type restriction.

Another optimization we'd like here is to have a run-length encoded representation (for types that implement `Eq`).

Differential Revision: D81612446