[hyperactor_mesh] prepare support for direct addressing
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1027

This change prepares the existing allocators for direct addressing, introduced in D79761721. Direct addressing will be needed to support host meshes.

With direct addressing we can no longer pre-determine the proc id (because the address itself is allocated/bound by the bootstrap process). This requires a changed to the allocator protocol, wherein we do not yet know the proc id in Created state; this is only revealed when the proc enters Running state.

In order for the users of the allocators to track this state transition, we introduce a "create key" -- an identifier that is associated with a proc created but not yet running. We use this identifier only to track this relationship.

Even so, this required fairly deep surgery in the remote process allocator, as it performed most of its state tracking on the basis of proc ids.
ghstack-source-id: 307876726

Differential Revision: [D81261853](https://our.internmc.facebook.com/intern/diff/D81261853/)