[monarch] hyperactor: introduce `Host`
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1191

This will be used as the foundation for HostMeshes. The diff also fixes up prefix routing for direct procid references.

This module defines [`Host`], which represents all the procs running on a host.
The procs themselves are managed by an implementation of [`ProcManager`], which may,
for example, fork new processes for each proc, or spawn them in the same process
for testing purposes.

The primary purpose of a host is to manage the lifecycle of these procs, and to
serve as a single front-end for all the procs on a host, multiplexing network
channels.

## Channel muxing

A [`Host`] maintains a single frontend address, through which all procs are accessible
through direct addressing: the id of each proc is the `ProcId::Direct(frontend_addr, proc_name)`.
In the following, the frontend address is denoted by `*`. The host listens on `*` and
multiplexes messages based on the proc name. When spawning procs, the host maintains
backend channels with separate addresses. In the diagram `#` is the backend address of
the host, while `#n` is the backend address for proc *n*. The host forwards messages
to the appropriate backend channel, while procs forward messages to the host backend
channel at `#`.

```text
                     ┌────────────┐
                 ┌───▶  proc *,1  │
                 │ #1└────────────┘
                 │
 ┌──────────┐    │   ┌────────────┐
 │   Host   │◀───┼───▶  proc *,2  │
*└──────────┘#   │ #2└────────────┘
                 │
                 │   ┌────────────┐
                 └───▶  proc *,3  │
                   #3└────────────┘
```
ghstack-source-id: 309489008

Differential Revision: [D82311596](https://our.internmc.facebook.com/intern/diff/D82311596/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D82311596/)!