Add supervision_events polling API to v1::ActorMeshRef
Summary:
Part of https://github.com/meta-pytorch/monarch/issues/1209

Add a `supervision_events` API for v1::ProcMeshRef and v1::ActorMeshRef to get the
list of supervision events that have accumulated on the ProcMeshAgent.

This needs to be configured so that supervision events don't crash the proc, and instead
are buffered into the mesh agent.

Reviewed By: mariusae

Differential Revision: D82131957