Endpoint call latency measurements
Summary:
Instrumenting the endpoint calls. This gives us latency measurements, along with the number of actors and name of the method. 

RFC: Is there a way to also get the name of the actor which whose endpoint we called?

Also fixing the python histogram record method to account for attributes.

Differential Revision: D82492042