Monarch web-crawler example. (#1359)
Summary:

This web-crawler example takes a target site and crawls it to a depth of 3.  It is initialized by adding the root to a central/single queue actor, and a number of parallel crawling actors take off the queue, crawl, and add to the queue, until the depth is exhausted.  The crawlers only crawl sub-pages of the base URL (just for safety/simplicity), and stop at the target depth.

Differential Revision: D83214162