Envvar for switching hyperactor server PEM (#1396)
Summary:

Right now, the hyperactor server PEM is hardcoded to `/var/facebook/x509_identities/server.pem`, but we have systems that don't have the right server.pem there.

This diff adds a new envvar that can be set for overriding it. It's feels a bit clunky configuring this by an envvar, but it would be a bigger lift to move these things to some other configeration.

Reviewed By: mariusae

Differential Revision: D83599004