Assign seq for PortHandle/PortRef (#1387)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1387

This diff is part of the effect to adding sequencing logic to sender actor. See D83371710 for details.

This diff adds the sequencing assignment capability to `PortHandle/PortRef`'s `send` methods. As a results, all methods sent from them will contain the sequence number header. It is possible the destination is not an actor, just a plain port receiver. That is okay. The sequence header will just be unused.

Reviewed By: shayne-fletcher, mariusae

Differential Revision: D83591305