Add supervision_event monitor to v1 ActorMeshRef (#1454)
Summary:

Part of https://github.com/meta-pytorch/monarch/issues/1209

Implement `supervision_event` for the v1 PythonActorMesh and References, for parity
with the v0 API.
This does not yet implement the goal of calling `__supervise__` on the owning PythonActor,
but that'll be the next step.

For parity with v0, had to add a few things:
* For the "monitor", poll the `actor_states` API every few seconds. If any Actor has
collected supervision events, surface those as a SupervisionError to python.
This should be optimized into a push-based API instead, or one based on accumulation
since we only need the first event.
* Before doing a `cast`, have to check if any previous cast had encountered a
supervision error, to avoid sending an undeliverable message
* Disallow spawning new actors on a ProcMeshAgent which has had a supervision event
in its past
* If there's an error that kills the whole proc (process exit, signal, etc.), the agent itself will
be down. Augment the `actor_states` API to handle this by reporting stopped for all remaining
ranks.

Reviewed By: colin2328, pablorfb-meta

Differential Revision: D82494380