[monarch] Plumb global config through from client to host
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1521

This diff adds new global config layer `ClientOverride`. When a new host mesh is allocated or created via simple bootstrap, the current value of the global config on the client is sent to each host mesh agent and set on the agent's `BootstrapProcManager`. When the `BootstrapProcManager` spawns a new process, the config snapshot is set in that process's global config at the `ClientOverride` layer, which is lower priority than any other layer (to account for, say, the user explicitly setting env vars on their hosts).

I plumbed this through a separate path than `BootstrapCommand` so that it's applicable for both the proc mesh trampoline approach and the simple bootstrap approach.
ghstack-source-id: 316029210
@exported-using-ghexport

Differential Revision: [D84578137](https://our.internmc.facebook.com/intern/diff/D84578137/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D84578137/)!