Only message PythonActorPanicWatcher when panic (#1540)
Summary:

`PythonActorPanicWatcher` only does anything useful (Returns an error in a message handler triggering a supervision error) when the value it is sent is a `SerializablePyErr`. When an endpoint succeeds, a `()` arrives in this channel, and the only thing we do is send a message to ourselves to re-enter the handler for `HandlePanic` and listen again for panics. You can think of it as a branch in a loop that just `continue`s. There are no side-effects or anything like that. Most messages are expected to succeed so this Actor will repeatedly message itself while doing nothing useful. 

If we change the channel to only send through `SerializablePyErr`, we still get the same effect where we surface endpoint panics as supervision events, this Actor can avoid sending itself a bunch of useless messages, and `handle_async_endpoint_panic` only needs to send a message through the channel when there is actually an error.

In fact this will actually allow panics to surface to the user quicker since the actor does not need to process every single successful message and can immediately surface the supervision event

I also took the liberty of renaming it because along the way of figuring out how everything works, I came to realize that this Actor is responsible for exactly one thing so I would like future maintainers to not have to think

Reviewed By: moonli

Differential Revision: D84639624