Fix infinite wait in actor mesh stop (#1598)
Summary:

In v1::ActorMesh::stop, there was occasionally an infinite wait for an actor to transition
to the "Stopped" state. I determined that sometimes an actor would get hung waiting
for a child actor to signal their parent that they had stopped.

Add a timeout for this check, so that in the case where the children are unresponsive,
the parent can still stop themselves. In that case the child actor becomes orphaned and might
never stop, but at least the parent can stop successfully and isn't left hanging.

Reviewed By: shayne-fletcher

Differential Revision: D84861696