Fix issues with two allocator tests (#1619)
Summary:

Didn't get around to re-enabling CI for these two tests yet, so a change in bootstrap behavior had broken them. This fixes that. 

Additionally, one of the late stage refactors broke the remoteprocess invocation, where instead of `tcp!::1:0` it needed to be `tcp![::1]:0`

Fixing the tests for the Github CI tests in a follow up, as it needs both setting an ipv4 binding fallback, as well as using the new attribute for setting binding to unspecified.

Reviewed By: shayne-fletcher

Differential Revision: D85049463