Fix supervision propagation (#1629)
Summary:

Part of https://github.com/meta-pytorch/monarch/issues/1209

Supervision events weren't being propagated any further than the first owning actor.
This is not intentional, we want to propagate them all the way to the client if they aren't handled.

The easiest way to do this is to return an Err variant from "handle", this will naturally become its
own supervision event and go to the actor's owner. This solution only works for PythonActors, and
will reach the client only through an exception coming from some endpoint.

Further work should be done to ensure the client crashes even if there are no active endpoints,
or even if the actor mesh which generated the exception is not reachable anymore.

Reviewed By: mariusae

Differential Revision: D85117996