Move supervision polling task to PythonActorMeshImpl constructor (#1628)
Summary:

Part of https://github.com/meta-pytorch/monarch/issues/1209

Supervision was not working in cases where there was no "await" of an endpoint.
This didn't provide error messages when an endpoint was broadcasted to without waiting
for a result. This can happen with "broadcast", but also with implicit messages like the `__init__`
constructor as well.

To fix this, move the creation of the state polling loop from the first call to "supervision_event" to closer
to construction time. It cannot be done in the constructor due to needing an "Instance" which is
not availble, particularly when unpickling an ActorMeshRef. Each "supervision_event" call clones
a tokio::watch::Receiver so it gets the same most recent event as every other endpoint.

This also allows a mesh to report an error to its owner much earlier if that mesh was idle.
Also improve the stringification of MeshFailure to not use the Debug format which is too verbose
for ActorSupervisionEvent.

Reviewed By: mariusae

Differential Revision: D85093139