Hook up pytest results for publishing on each Github Action (#1648)
Summary:

The pytorch test-infra jobs already support submitting test reports for display on a job.
This can improve the user experience when there's a problem with the CI.

Right now this is just for python tests, but cargo also supports the JUnit xml format, so we
can add that later too.

Differential Revision: D85294258