Downgrade message log level to TRACE
Summary:
We found this log not helpful when debugging S576170. But it causes log spew and makes it hard to find other useful logs.

Downgrade to TRACE instead of deleting it just in case we might need it in the future debugging (at that time we can manually bump its level).

Reviewed By: shayne-fletcher

Differential Revision: D85533659