Migrator monarch/python/monarch/_src/actor to pyre-strict
Summary:
pyre had a couple of gaps for how we handled endpoints that interfered with full turning typing on that are mostly addressed with pyrefly. 

This is part 1 of a series of diffs to turn on pyre-strict, leaning someone generously on Any's for now, which then will be followed up by enabling pyrefly.

Differential Revision: D85356082