Add  python bindings for serve/dial
We are going to end up with cases where during bootstrap, we will need to make some initial connections in order to discover the information needed to call attach_to_workers. For instance, in torchx bootstrap, the workers know where the client is, but not vice-versa.  While we could use a non-monarch tcp server/client in the bootstrap function, we would have to re-implement a bunch of the connection properties (dial before serve, tls handling). So the most flexible thing seems to be just to expose dial/serve on our channels and then let the bootstrap code use it to get the initial information.

Differential Revision: [D86321192](https://our.internmc.facebook.com/intern/diff/D86321192/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D86321192/)!

[ghstack-poisoned]