[Logging] put the actor name in the python logger
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1733

Injects [actor=<the fully qualified actor name>] into the log message of the python actor.

The prefix is configurable by redefining `monarch.actor.per_actor_logging_prefix` callback. Setting to None disables the filter that alters the logs.

Hopefully this works pretty broadly since it hooks in as a filter on logging handlers.
ghstack-source-id: 321417368
@exported-using-ghexport

Differential Revision: [D85994688](https://our.internmc.facebook.com/intern/diff/D85994688/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D85994688/)!