[monarch] Add message envelope header for no return
Sometimes when we send a message, we want it to be fully fire-and-forget,
including if the destination is not even reachable. This is typically only used in
scenarios like:
* When shutting down the system, we try to ask a process to nicely shut itself down
before ungracefully killing it. If the message is undeliverable, we can just proceed with
killing the process (it's probably already dead anyways)
* Replying to a message. If the sender is down, there's nothing the current actor can do about it

This should be used sparingly as it could hide real errors, like your messages not getting sent.

Add a header to MessageEnvelope for this use case, which avoids posting the Undelivered
message back to the sender.
Use it with a send of the StopAll message.

Differential Revision: [D86315780](https://our.internmc.facebook.com/intern/diff/D86315780/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D86315780/)!

[ghstack-poisoned]